change permissions

+++ conf/converter/compareHashes.sh
@@ -2,6 +2,11 @@
 
 ctph_file=$1
 
+if [ ! -f $ctph_file ]
+then
+	exit 1
+fi
+
 echo 'ssdeep,1.1--blocksize:hash:hash,filename' | cat - "$ctph_file" > temp && mv temp "$ctph_file"
 
 mkdir -p ../../hashes
